avoid unnecessary recalculations of required height
Calculating the required height is an expensive operation.
`Element.getComputedStyles()`, `Element.scrollHeight` and other features
used require the browser to recalculate styles and layout. We should make
sure that this is only done once per render.

Such an resize could be triggered by both the input event and a changed
argument in the same render. Actually that's very likely as could be seen
by this template:

  <textarea
    value={{this.val}}
    {{autoresize this.val}}
    {{on "input" (action (mut this.val) value="target.value")}}
  />

This requires to track if resize has been scheduled even if arguments are
changed. This is something which is not good supported by functional
modifiers. Therefore the implementation is refactored to use class modifiers
instead as part of this change.